initial sprint 2
added :
- splash screen
- customer icon at appbar
