+++ lib/app/routes.dart
@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:vetclinic/ui/view/forgot_password_view.dart';
-import 'package:vetclinic/ui/view/vet_home_view.dart';
-import 'package:vetclinic/ui/view/login_view.dart';
-import 'package:vetclinic/ui/view/profile_view.dart';
-import 'package:vetclinic/ui/view/register_view.dart';
+import 'package:vetclinic/ui/screen/customer_home/customer_home_view.dart';
+import 'package:vetclinic/ui/screen/forgot_password/forgot_password_view.dart';
+import 'package:vetclinic/ui/screen/history/history_view.dart';
+import 'package:vetclinic/ui/screen/login/login_view.dart';
+import 'package:vetclinic/ui/screen/profile/profile_view.dart';
+import 'package:vetclinic/ui/screen/register/register_view.dart';
+import 'package:vetclinic/ui/screen/vet_home/vet_home_view.dart';
 
-import '../ui/view/customer_home_view.dart';
 
 class Routes {
   static const String customerHomeRoute = '/customerHome';
@@ -14,6 +15,7 @@ class Routes {
   static const String registerRoute = '/register';
   static const String forgotPassRoute = '/forgotPass';
   static const String profileRoute = '/profile';
+  static const String historyRoute = '/history';
 
   static Route<dynamic>? createRoute(settings) {
     switch (settings.name) {
@@ -34,6 +36,9 @@ class Routes {
       
       case profileRoute :
         return ProfileView.route();
+
+      case historyRoute :
+        return HistoryView.route();
       
     }
     return null;
