+++ lib/ui/screen/forgot_password/forgot_password_viewmodel.dart
@@ -1,7 +1,7 @@
 import 'package:map_mvvm/map_mvvm.dart';
 
-import '../app/app.dart';
-import '../services/firebase/firebase_service.dart';
+import '../../../app/app.dart';
+import '../../../services/firebase/firebase_service.dart';
 
 class ForgotPasswordViewModel extends Viewmodel {
   FirebaseService get _service => locator<FirebaseService>();
