+++ lib/ui/screen/profile/profile_viewmodel.dart
@@ -67,9 +67,9 @@
 import 'package:map_mvvm/failure.dart';
 import 'package:map_mvvm/viewmodel.dart';
 
-import '../../../app/service_locator.dart';
-import '../services/firebase/firebase_service.dart';
-import '../model/Users.dart';
+import '../../../../../app/service_locator.dart';
+import '../../../services/firebase/firebase_service.dart';
+import '../../../model/Users.dart';
 
 class ProfileViewModel extends Viewmodel {
   FirebaseService get _service => locator<FirebaseService>();
