+++ lib/ui/screen/register/register_viewmodel.dart
@@ -80,8 +80,8 @@
 import 'package:map_mvvm/failure.dart';
 import 'package:map_mvvm/viewmodel.dart';
 
-import '../../../app/service_locator.dart';
-import '../services/firebase/firebase_service.dart';
+import '../../../../../app/service_locator.dart';
+import '../../../services/firebase/firebase_service.dart';
 
 
 class RegisterViewModel extends Viewmodel {
