+++ web/splash/splash.js
@@ -1,0 +1,7 @@
+function removeSplashFromWeb() {
+  const elem = document.getElementById("splash");
+  if (elem) {
+    elem.remove();
+  }
+  document.body.style.background = "transparent";
+}
